Extracts signals to separate file
