+++ djangoserver/djangoserver/settings.py
@@ -49,7 +49,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
     'users.apps.UsersConfig',
-    'forum.apps.ForumConfig',
+    'forum',
 ]
 
 MIDDLEWARE = [
