+++ djangoserver/forum/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class ForumConfig(AppConfig):
     name = 'forum'
+
+    def ready(self):
+        import forum.signals
